move IconService and Provider up again
