--- src/modules/iconSelect/injection/index.js
@@ -1,5 +1,0 @@
-import { IconService } from '../services/IconService';
-
-export const iconSelectModule = ($injector) => {
-	$injector.registerSingleton('IconService', new IconService());
-};
