+++ src/modules/map/components/olMap/handler/draw/OlDrawHandler.js
@@ -49,7 +49,7 @@ const defaultStyleOption = {
 export class OlDrawHandler extends OlLayerHandler {
 	constructor() {
 		super(DRAW_LAYER_ID);
-		const { TranslationService, MapService, EnvironmentService, StoreService, GeoResourceService, FileStorageService, OverlayService, StyleService, MeasurementStorageService } = $injector.inject('TranslationService', 'MapService', 'EnvironmentService', 'StoreService', 'GeoResourceService', 'FileStorageService', 'OverlayService', 'StyleService', 'MeasurementStorageService');
+		const { TranslationService, MapService, EnvironmentService, StoreService, GeoResourceService, FileStorageService, OverlayService, StyleService, MeasurementStorageService, IconService } = $injector.inject('TranslationService', 'MapService', 'EnvironmentService', 'StoreService', 'GeoResourceService', 'FileStorageService', 'OverlayService', 'StyleService', 'MeasurementStorageService', 'IconService');
 		this._translationService = TranslationService;
 		this._mapService = MapService;
 		this._environmentService = EnvironmentService;
@@ -59,6 +59,7 @@ export class OlDrawHandler extends OlLayerHandler {
 		this._overlayService = OverlayService;
 		this._styleService = StyleService;
 		this._storageHandler = MeasurementStorageService;
+		this._iconService = IconService;
 
 		this._vectorLayer = null;
 		this._draw = null;
@@ -502,7 +503,7 @@ export class OlDrawHandler extends OlLayerHandler {
 
 	_getStyleOption() {
 		if (this._storeService.getStore().getState().draw.style == null) {
-			setStyle(defaultStyleOption);
+			setStyle({ ...defaultStyleOption, symbolSrc: this._iconService.default().toBase64() });
 		}
 		return this._storeService.getStore().getState().draw.style;
 	}
